+++ src/proto.h
@@ -178,6 +178,7 @@ void qsort(void *base, size_t elm_count, size_t elm_size, int (*cmp)(const void 
 # include "normal.pro"
 # include "ops.pro"
 # include "option.pro"
+# include "optionstr.pro"
 # include "popupmnu.pro"
 # if defined(FEAT_PROFILE) || defined(FEAT_RELTIME)
 # include "profiler.pro"
